requirements.txt
opencv-python~=4.7.0.72
numpy~=1.23.5
pillow~=9.5.0